refactor: combine custom types
